refactor(components): share one context-menu implementation
ContextMenuItem, ContextMenuSeparator and ContextMenuSubmenu existed as
byte-identical copies under content/, tab/ and sidebar/, and the viewport
clamping helpers lived only in sidebar/ even though every menu needs
them. A change to how a menu item renders had to be found and applied in
each copy by hand, and only the sidebar menu stayed on screen near a
viewport edge.

Move them into components/context_menu/ so a menu change lands in one
place. The sidebar copy declared `label: &'static str`; the shared
component takes anything convertible into String, which accepts every
existing call site unchanged. Rendered markup is identical.

The eleven clamping tests move with the code they cover; the two tests
that exercise sidebar-specific behavior stay behind.
